test: organise + 2 new tests (#38)
